fix: v1.35 全维度审计修复 — 安全/代码/数据库/文档全面加固
P0 修复：数据库Schema不匹配(3处)、权限缺失(TopicController)、
  SQL日志泄露、CORS通配符、Refresh Token类型隔离、OOM风险、
  N+1查询(3处)、Docker构建失败、@Async未生效、文档表数矛盾(4处)

P1 修复：安全响应头、登录锁定竞态、私信XSS过滤、分页参数校验、
  数据库索引优化、索引命名规范、API文档路径修正、组件计数统一

+++ src/main/java/com/example/edu_project/controller/ReportController.java
@@ -49,8 +49,8 @@ public class ReportController {
     @Operation(summary = "获取我的举报记录")
     @GetMapping("/my")
     public Result<IPage<ReportVO>> getMyReports(
-            @RequestParam(defaultValue = "1") @Min(1) Long page,
-            @RequestParam(defaultValue = "10") @Min(1) @Max(100) Long pageSize) {
+            @RequestParam(defaultValue = "1") @Min(1) Integer page,
+            @RequestParam(defaultValue = "10") @Min(1) @Max(100) Integer pageSize) {
         Long userId = SecurityUtils.getCurrentUserIdOrNull();
         if (userId == null) {
             throw new BusinessException(401, "请先登录");
